+++ internal/api/router.go
@@ -31,6 +31,7 @@ func RegisterRouter(e *echo.Echo, repo port.Repository, cfg *config.Config, mtxC
 
 	handler.RegisterStream(api, d)
 	handler.RegisterAsset(api, d)
+	handler.RegisterOperator(api, d)
 	handler.RegisterAlgorithm(api, d)
 	handler.RegisterAlgorithmBinding(api, d)
 	handler.RegisterRecord(api, d)